# Slightly speed up scripts