@nik9000 2021-02-03
The actual speedup from this is pretty disappointing. The async profiler has betrayed me! But its a very very easy change so I figure its worth it.