@nik9000 2021-02-03
Benchmarks just finished:

Before:
```
Benchmark                       (indexingBufferMb)       (script)  Mode  Cnt   Score   Error  Units
ScriptScoreBenchmark.benchmark                  16     expression  avgt   10  63.950 ± 1.444  ns/op
ScriptScoreBenchmark.benchmark                  16          metal  avgt   10  52.474 ± 0.572  ns/op
ScriptScoreBenchmark.benchmark                  16  painless_cast  avgt   10  71.872 ± 0.472  ns/op
ScriptScoreBenchmark.benchmark                  16   painless_def  avgt   10  72.232 ± 1.007  ns/op
```
After:
```
Benchmark                       (indexingBufferMb)       (script)  Mode  Cnt   Score   Error  Units
ScriptScoreBenchmark.benchmark                  16     expression  avgt   10  63.203 ± 0.063  ns/op
ScriptScoreBenchmark.benchmark                  16          metal  avgt   10  52.388 ± 0.312  ns/op
ScriptScoreBenchmark.benchmark                  16  painless_cast  avgt   10  70.971 ± 0.235  ns/op
ScriptScoreBenchmark.benchmark                  16   painless_def  avgt   10  70.097 ± 0.469  ns/op
```

Like I said - its a small improvement for painless scripts. But its a small change.