@nik9000 2021-02-03
The async profiler led me to believe it'd be a bigger improvement or else I wouldn't even have looked. But still.